requirements.txt
django==4.2.5
djangorestframework==3.14.0
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
textblob